+++ test/byContract/IexecPocoBoost/IexecPocoBoostDelegate.test.ts
@@ -57,6 +57,7 @@ const EIP712DOMAIN_SEPARATOR = 'EIP712DOMAIN_SEPARATOR';
 const BALANCES = 'm_balances';
 const FROZENS = 'm_frozens';
 const WORKERPOOL_STAKE_RATIO = 30;
+const kittyAddress = '0x99c2268479b93fDe36232351229815DF80837e23';
 const { domain, domainSeparator } = buildDomain();
 const appPrice = 1000;
 const datasetPrice = 1_000_000;
@@ -1332,11 +1333,20 @@ describe('IexecPocoBoostDelegate', function () {
             const initialAppOwnerBalance = 5;
             const initialDatasetOwnerBalance = 6;
             const initialSchedulerBalance = 7;
+            const initialKitty = 10_000_000_010; // MIN_KITTY * 10 + 10,
             const schedulerDealStake = computeSchedulerDealStake(workerpoolPrice, volume);
             const schedulerTaskStake = schedulerDealStake / volume;
+            // Setup: MIN_REWARD < reward < available
+            // Further assertion on scheduler kitty reward will fail if the
+            // KITTY_RATIO constant is someday updated in the source code.
+            const expectedSchedulerKittyRewardForTask1 =
+                (initialKitty * // total kitty
+                    10) / // KITTY_RATIO
+                100; // percentage
+
             await iexecPocoBoostInstance.setVariables({
                 [BALANCES]: {
-                    [iexecPocoBoostInstance.address]: initialIexecPocoBalance,
+                    [iexecPocoBoostInstance.address]: initialIexecPocoBalance + initialKitty,
                     [requester.address]: initialRequesterBalance + dealPrice,
                     [worker.address]: initialWorkerBalance,
                     [scheduler.address]: initialSchedulerBalance + schedulerDealStake,
@@ -1345,6 +1355,7 @@ describe('IexecPocoBoostDelegate', function () {
                 },
                 [FROZENS]: {
                     [requester.address]: initialRequesterFrozen,
+                    [kittyAddress]: initialKitty,
                 },
             });
             await signOrders(domain, orders, ordersActors);
@@ -1379,7 +1390,7 @@ describe('IexecPocoBoostDelegate', function () {
             await expectBalance(
                 iexecPocoBoostInstance,
                 iexecPocoBoostInstance.address,
-                initialIexecPocoBalance + dealPrice + schedulerDealStake,
+                initialIexecPocoBalance + dealPrice + schedulerDealStake + initialKitty,
             );
             // Check requester balance and frozen
             await expectBalance(iexecPocoBoostInstance, requester.address, initialRequesterBalance);
@@ -1393,6 +1404,8 @@ describe('IexecPocoBoostDelegate', function () {
             // Check scheduler balance and frozen
             await expectBalance(iexecPocoBoostInstance, scheduler.address, initialSchedulerBalance);
             await expectFrozen(iexecPocoBoostInstance, scheduler.address, schedulerDealStake);
+            // Check kitty frozen
+            await expectFrozen(iexecPocoBoostInstance, kittyAddress, initialKitty);
             // Check app provider balance
             await expectBalance(
                 iexecPocoBoostInstance,
@@ -1410,7 +1423,9 @@ describe('IexecPocoBoostDelegate', function () {
             // we just need to verify here that some worker reward value will be
             // transferred
             expect(expectedWorkerReward).to.be.greaterThan(0);
-            const expectedSchedulerReward = workerpoolPrice - expectedWorkerReward;
+            const expectedSchedulerBaseReward = workerpoolPrice - expectedWorkerReward;
+            const expectedSchedulerFullReward =
+                expectedSchedulerBaseReward + expectedSchedulerKittyRewardForTask1;
 
             await expect(
                 iexecPocoBoostInstance
@@ -1426,7 +1441,7 @@ describe('IexecPocoBoostDelegate', function () {
                     ),
             )
                 .to.emit(iexecPocoBoostInstance, 'Seize')
-                .withArgs(requester.address, expectedWorkerReward + appPrice + datasetPrice, taskId) //TODO: Seize app + dataset + workerpool price
+                .withArgs(requester.address, taskPrice, taskId)
                 .to.emit(iexecPocoBoostInstance, 'Transfer')
                 .withArgs(iexecPocoBoostInstance.address, worker.address, expectedWorkerReward)
                 .to.emit(iexecPocoBoostInstance, 'Reward')
@@ -1443,6 +1458,16 @@ describe('IexecPocoBoostDelegate', function () {
                 .withArgs(iexecPocoBoostInstance.address, scheduler.address, schedulerTaskStake)
                 .to.emit(iexecPocoBoostInstance, 'Unlock')
                 .withArgs(scheduler.address, schedulerTaskStake)
+                .to.emit(iexecPocoBoostInstance, 'Seize')
+                .withArgs(kittyAddress, expectedSchedulerKittyRewardForTask1, taskId)
+                .to.emit(iexecPocoBoostInstance, 'Transfer')
+                .withArgs(
+                    iexecPocoBoostInstance.address,
+                    scheduler.address,
+                    expectedSchedulerFullReward,
+                )
+                .to.emit(iexecPocoBoostInstance, 'Reward')
+                .withArgs(scheduler.address, expectedSchedulerFullReward, taskId)
                 .to.emit(iexecPocoBoostInstance, 'ResultPushedBoost')
                 .withArgs(dealId, taskIndex, results);
             expect(oracleConsumerInstance.receiveResult).to.have.been.calledOnceWith(
@@ -1470,14 +1495,15 @@ describe('IexecPocoBoostDelegate', function () {
                 iexecPocoBoostInstance.address,
                 initialIexecPocoBalance +
                     (taskPrice + schedulerTaskStake) * remainingTasksToPush +
-                    expectedSchedulerReward, // TODO: Remove after scheduler reward feature
+                    initialKitty -
+                    expectedSchedulerKittyRewardForTask1,
             );
             // Check requester balance and frozen
             await expectBalance(iexecPocoBoostInstance, requester.address, initialRequesterBalance);
             await expectFrozen(
                 iexecPocoBoostInstance,
                 requester.address,
-                initialRequesterFrozen + taskPrice * remainingTasksToPush + expectedSchedulerReward, // TODO: Remove after scheduler reward feature
+                initialRequesterFrozen + taskPrice * remainingTasksToPush,
             );
             // Check worker balance
             await expectBalance(
@@ -1501,12 +1527,18 @@ describe('IexecPocoBoostDelegate', function () {
             await expectBalance(
                 iexecPocoBoostInstance,
                 scheduler.address,
-                initialSchedulerBalance + schedulerTaskStake,
+                initialSchedulerBalance + schedulerTaskStake + expectedSchedulerFullReward,
             );
             await expectFrozen(
                 iexecPocoBoostInstance,
                 scheduler.address,
                 schedulerTaskStake * remainingTasksToPush,
+            );
+            // Check kitty frozen
+            await expectFrozen(
+                iexecPocoBoostInstance,
+                kittyAddress,
+                initialKitty - expectedSchedulerKittyRewardForTask1,
             );
         });